refactor: インポート文の統一と中間層削除
- packages/shared/src/index.ts を更新
  - 新しい型定義と型ガード関数のエクスポートを追加
  - filesystem.ts, route-params.ts の型をエクスポート
- packages/backend/src/routes/files.ts を更新
  - ProjectFilesParams をsharedから直接インポート
- packages/backend/src/services/fileSystemService.ts を更新
  - DirectoryPermissions, FileSystemError をsharedから直接インポート
  - ローカルのFileSystemErrorクラス定義を削除
- packages/frontend/src/components/custom-ui/ を更新
  - file-tree.ts から components.ts への直接インポートに変更
- packages/frontend/src/types/file-tree.ts を削除
  - 不要な中間層ファイルを削除

要件 4.1, 4.2, 5.1, 5.2 に対応
