--- packages/frontend/src/types/file-tree.ts
@@ -1,7 +1,0 @@
-/**
- * ファイルツリーに関する型定義
- *
- * @deprecated このファイルは非推奨です。components.tsを使用してください。
- */
-
-export type { FileTreeProps, FileTreeItemProps } from './components.js';
